\name{summary.sarprobit}
\alias{summary.sarprobit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print the results of the SAR probit estimation via MCMC}
\description{
Print the results of the SAR probit estimation via MCMC
}
\usage{
\method{summary}{sarprobit}(object, var_names = NULL, file = NULL, ...)
}
\arguments{
  \item{object}{structure with the results from sar_probit_mcmc}
  \item{var_names}{vector with names for the parameters under analysis}
  \item{file}{TO BE DONE}
  \item{...}{Additional arguments}
}
\value{
This functions does not return any values.
}
\author{
Miguel Godinho de Matos <miguelgodinhomatos@cmu.edu>
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
"summary.sarprobit"
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
